Fixes for 32bit, and chromebook images, and a bit of documentation

+++ chromebook-arm-hp.sh
@@ -320,8 +320,12 @@ rm -rf ${basedir}/kernel ${basedir}/bootp ${basedir}/root ${basedir}/kali-$archi
 
 echo "Generating sha1sum for kali-$1-spring.img"
 sha1sum kali-$1-spring.img > ${basedir}/kali-$1-spring.img.sha1sum
+# Don't pixz on 32bit, there isn't enough memory to compress the images.
+MACHINE_TYPE=`uname -m`
+if [ ${MACHINE_TYPE} == 'x86_64' ]; then
 echo "Compressing kali-$1-spring.img"
 pixz ${basedir}/kali-$1-spring.img ${basedir}/kali-$1-spring.img.xz
 rm ${basedir}/kali-$1-spring.img
 echo "Generating sha1sum for kali-$1-spring.img.xz"
 sha1sum kali-$1-spring.img.xz > ${basedir}/kali-$1-spring.img.xz.sha1sum
+fi
